Enterprise: add schools / Login: native hash algorithm
In the enterprise the (main)admin can now add schools to the system.
Some refinement is yet to be done.

The used login-hash algorithm gave problems from time to time (likely
caused by the salt). Now using the native password_hash and
password_verify functions.
-> fix problem + more secure over time

NOTE: the password-fields in the database should have a length of 255
characters (to allow future hashes to be stored completely. Currently
only 60 characters are used)